feat: change password moved to separate file
